Get the kelp status for a particular year

Status is calculated by dividing the given year's kelp area by the historical median across 1984-2013 for each kelp segment.

Usage

get_kelp_status(
  annual_time_series,
  kelp_presence,
  ...,
  status_year = NULL,
  segment_id = NULL
)

Arguments

annual_time_series

Annual time series generated by extract_time_series(). Does not work with quarterly time series

kelp_presence

Kelp presence data frame generated by get_kelp_presence()

...

unused

status_year

Year to calculate the status for, or NULL for the latest year in the time series. Must be later than 2013

segment_id

A character vector containing the subset of segment_id's of interest, e.g. c("CA_71", "CA_78"), or NULL for all segment IDs (default).

Value

A data frame with statuses for each Segment_ID in annual_time_series. A value of -999 indicates that there is not enough data to determine a status for the given kelp segment